Remove adapter_method from DatabaseConfig
It feels a bit off for a configuration object to be giving out details
like method names that should be called on `Base`. This commit removes the
`adapter_method` method from `DatabaseConfig` in favor of an
`adapter_method_for` method directly on `ConnectionAdapters`.

While I was in here, I also split out the adapter loading behavior from
`resolve_pool_config` and moved it as well.