Pass role around to connection methods
This is gross because we're not using kwargs and I think we may want to
move to a better method signature but right now we're just trying to see
if this works.

The change here pushes role around in the connection methods like shards
does. The goal is to treat roles and shards more similarly and get rid
of the handlers for handling the roles.

---

Some notes:

Retrieve connection has access to:
  spec name
  pool key
  connection handler

If we dropped the reading connection handler and had a single one, how do we
lookup connections

We'll need to pass a role and pool key

spec_name => pool_manager
  pool_manager.pool_configs[:default][:writing]
  pool_manager.pool_configs[:default][:reading]
  { default => { writing => pool_config, reading => pool_config } }

=> PoolManager(schema_cache, db_config)
=> PoolConfig(db_config, pool, spec_name)

- should pool config also store role and shard keys? or do we just
put those in a ma[?
- shards have the same schema cache (right?) so we can push that up to pool_manager
  - writing/reading also has the same schema cache
  - veritcal sharding doesn't have the same schema cache but those are all

what we need to do this:
- define current_role
- pass current_role around in these methods
- create a new pool manager class
- update connects_to/establish_connection
- deprecate the old way if applicable
- remove deprecated code

considerations:
- if changes are too great, add deprecation and optional opt-in
- what are the consequences of more granular connection swapping?
- are there other patterns that can achieve this?
- are there performance issues in this new approach?
- callbacks are tricky, need to think through this