Ok - the state of things
Things pass except for the fixture connection tests and one test that
is getting too many shards in the list because we're no longer cleaning
up connections at the end of the tests.

We need a way I think to turn off shared connection pools for the AR
tests, or at least when we're not doing multi-connection work.

https://github.com/rails/rails/pull/40384 is interesting but I didn't
have time to make it work with our stuff.

So basically we're in a better state but I'm not quite sure yet how to
get just the 2 tests that need shared pools to use them. Thoughts?